:hammer: Implement Screen navigation and switching in base class for multi-platform reuse
